@@ -5,40 +5,40 @@
     "keywords": [],
     "license": "MIT",
     "require": {
-        "php": "^7.4|^8.1",
+        "php": "^8.1",
         "blade-ui-kit/blade-icons": "^1.1",
-        "fakerphp/faker": "^1.21",
-        "filament/forms": "^2.0",
-        "filament/tables": "^2.7",
-        "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^7.0.1",
-        "intervention/image": "^2.7",
+        "fakerphp/faker": "^1.9.1",
+        "filament/forms": "^3.0-stable",
+        "filament/notifications": "^3.0-stable",
+        "filament/tables": "^3.0-stable",
+        "guzzlehttp/guzzle": "^7.2",
         "knuckleswtf/scribe": "^4.25",
-        "laravel/framework": "^9.6",
-        "laravel/sanctum": "^2.11",
-        "laravel/tinker": "^2.5",
-        "livewire/livewire": "^2.7",
+        "laravel/framework": "^10.10",
+        "laravel/sanctum": "^3.3",
+        "laravel/tinker": "^2.8",
+        "livewire/livewire": "^3.2",
         "maatwebsite/excel": "^3.1",
         "rap2hpoutre/laravel-log-viewer": "^2.2",
-        "rappasoft/laravel-livewire-tables": "^1.20",
+        "rappasoft/laravel-livewire-tables": "^3.1",
         "spatie/browsershot": "^3.52",
         "spatie/laravel-activitylog": "^4.4",
-        "spatie/laravel-ignition": "^1.6",
+        "spatie/laravel-ignition": "^2.0",
         "spatie/laravel-model-states": "^2.1",
         "spatie/laravel-permission": "^5.3",
-        "spatie/laravel-welcome-notification": "^2.2",
-        "studio/laravel-totem": "^9.0"
+        "spatie/laravel-welcome-notification": "^2.2"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.3",
         "laravel/breeze": "^1.4",
-        "laravel/sail": "^1.0.1",
+        "laravel/pint": "^1.0",
+        "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^6.10",
-        "pestphp/pest": "^1.21",
-        "pestphp/pest-plugin-laravel": "^1.2",
-        "pestphp/pest-plugin-livewire": "^1.0",
-        "phpunit/phpunit": "^9.5.10"
+        "nunomaduro/collision": "^7.0",
+        "pestphp/pest": "^2.26",
+        "pestphp/pest-plugin-faker": "^2.0",
+        "pestphp/pest-plugin-laravel": "^2.2",
+        "phpunit/phpunit": "^10.0",
+        "spatie/laravel-ignition": "^2.0"
     },
     "autoload": {
         "psr-4": {
@@ -46,7 +46,9 @@
             "Database\\Factories\\": "database/factories/",
             "Database\\Seeders\\": "database/seeders/"
         },
-        "files": ["app/helper.php"]
+        "files": [
+            "app/helper.php"
+        ]
     },
     "autoload-dev": {
         "psr-4": {
@@ -56,7 +58,8 @@
     "scripts": {
         "post-autoload-dump": [
             "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
-            "@php artisan package:discover --ansi"
+            "@php artisan package:discover --ansi",
+            "@php artisan filament:upgrade"
         ],
         "post-update-cmd": [
             "@php artisan vendor:publish --tag=laravel-assets --ansi"
@@ -84,6 +87,6 @@
             "pestphp/pest-plugin": true
         }
     },
-    "minimum-stability": "dev",
+    "minimum-stability": "stable",
     "prefer-stable": true
 }
