@@ -2,40 +2,57 @@
 
 namespace App\Http\Livewire\Account;
 
+use Filament\Forms\Components\Section;
+use Filament\Forms\Components\Toggle;
 use Filament\Forms\Contracts\HasForms;
+use Filament\Forms\Form;
+use Filament\Notifications\Notification;
 use Livewire\Component;
 
 class OrderConfigForm extends Component implements HasForms
 {
     use \Filament\Forms\Concerns\InteractsWithForms;
 
-    public bool $is_for_evening = false;
+    public ?array $data = [];
+
 
     public function mount()
     {
-        $this->is_for_evening = auth()->user()->order_for_evening;
+        $this->data = [
+            'is_for_evening' => auth()->user()->order_for_evening,
+        ];
     }
 
-    public function getFormSchema(): array
+    public function form(Form $form): Form
     {
-        return [
-            \Filament\Forms\Components\Grid::make()
-                ->schema([
-                    \Filament\Forms\Components\Toggle::make('is_for_evening')
-                        ->label('Configurer automatiquement mes commandes pour le soir. Si vous ne cochez pas cette case, vos commandes seront configurées pour le midi.')
-                ]),
-        ];
+        return $form
+            ->schema([
+           Section::make('Préférences de commande')
+               ->aside()
+               ->description('Si vous cochez cette case, vous commanderez pour le soir. Sinon, vous commanderez pour le midi')
+               ->schema([
+                    Toggle::make('is_for_evening')
+                        ->label('Commander pour le soir')
+               ])
+
+           // ...
+            ])
+            ->statePath('data');
     }
 
     public function save()
     {
         $this->validate();
-
         auth()->user()->update([
-            'order_for_evening' => $this->is_for_evening,
+            'order_for_evening' => $this->data['is_for_evening'],
         ]);
 
-        session()->flash('success', 'Vos préférences ont été mises à jour !');
+        Notification::make()
+            ->title('Préférences de commande')
+            ->body('Vos préférences de commande ont été mises à jour avec succès.')
+            ->success()
+            ->send();
+
         return redirect()->route('profile');
     }
 
