@@ -1,3 +1,3 @@
 <div {{ $attributes }}>
-    <img class="h-auto w-3/5 mx-auto mt-3 mb-6" src="{{ asset('images/logo-dash.png') }}">
+    {{-- <img class="h-auto w-3/5 mx-auto mt-3 mb-6" src="{{ asset('images/logo-dash.png') }}"> --}}
 </div>
