@@ -1,6 +1,6 @@
 <x-app-layout>
     <x-section-header title="Historique des pointages"></x-section-header>
     <div class="bg-white px-6 py-4 rounded-md shadow-lg">
-        <livewire:orders.check-in-breakfast-table />
+        <livewire:tables.breakfast-completed-table />
     </div>
 </x-app-layout>
