@@ -17,6 +17,6 @@
         @endif
     </x-section-header>
     <div class="bg-white px-6 py-4 rounded-md shadow-lg">
-        <livewire:orders.orders-table />
+        <livewire:tables.order-table />
     </div>
 </x-app-layout>
