@@ -1,7 +1,6 @@
 <x-app-layout>
     <x-section-header title="Commandes hebdomadaires"> </x-section-header>
-
-    <div class="bg-white px-6 py-4 rounded-md shadow-lg">
-        <livewire:orders.orders-summary-table />
+    <div class="px-6 py-4">
+        <livewire:tables.weekly-order-table />
     </div>
 </x-app-layout>
