@@ -1,7 +1,6 @@
 <x-app-layout>
-    <x-section-header title="Commandes journalières"> </x-section-header>
-
-    <div class="bg-white px-6 py-4 rounded-md shadow-lg">
-        <livewire:orders.today-orders />
+    <x-section-header title="Commandes journalières" />
+    <div class="px-6 py-4">
+        <livewire:tables.dayli-reporting-table />
     </div>
 </x-app-layout>
