@@ -1,7 +1,6 @@
 <x-app-layout>
     <x-section-header title="Reporting petit dejeuner" />
-
-    <div class="bg-white px-6 py-4 rounded-md shadow-lg">
-        <livewire:reporting.check-in-breakfast-table />
+    <div class="px-6 py-4">
+        <livewire:tables.breakfast-reporting-table />
     </div>
 </x-app-layout>
