@@ -16,7 +16,7 @@
             </x-slot>
         @endcan
     </x-section-header>
-    <div class="bg-white px-6 py-4 rounded-md shadow-lg">
-        <livewire:suggestions.suggestion-box-table />
+    <div class=" px-6 py-4">
+        <livewire:tables.suggestion-table />
     </div>
 </x-app-layout>
