@@ -16,8 +16,7 @@
             </x-slot>
         @endcan
     </x-section-header>
-
-    <div class="bg-white px-6 py-4 rounded-md shadow-lg">
-        <livewire:user-types.user-types-table />
+    <div class="px-6 py-4">
+        <livewire:tables.user-type-table />
     </div>
 </x-app-layout>
