@@ -1,9 +1,7 @@
 import { defineConfig, loadEnv } from 'vite'
 import laravel, { refreshPaths } from 'laravel-vite-plugin'
 
-
 export default defineConfig(({ command, mode }) => {
-    //Load the env variables that are prefixed with VITE_
     const env = loadEnv(mode, process.cwd(), 'VITE_')
     return {
         plugins: [
@@ -11,8 +9,12 @@ export default defineConfig(({ command, mode }) => {
                 input: ['resources/css/app.css', 'resources/js/app.js'],
                 refresh: [
                     ...refreshPaths,
-                    'app/Http/Livewire/**',
-                    'app/Tables/Columns/**',
+                'app/Filament/**',
+                'app/Forms/Components/**',
+                'app/Livewire/**',
+                'app/Infolists/Components/**',
+                'app/Providers/Filament/**',
+                'app/Tables/Columns/**',
                 ],
             }),
         ],
